Refactored code, split eac4 and ghg configs
