+++ atmospheric_explorer/data_interface/__init__.py
@@ -1,0 +1,3 @@
+# pylint: disable=missing-module-docstring
+# ruff: noqa: F401
+from .cams_interface import CAMSDataInterface
