+++ atmospheric_explorer/data_interface/ghg/__init__.py
@@ -1,0 +1,4 @@
+# pylint: disable=missing-module-docstring
+# ruff: noqa: F401
+from .ghg import InversionOptimisedGreenhouseGas
+from .ghg_config import GHGConfig
