+++ atmospheric_explorer/ui/session_state.py
@@ -31,7 +31,8 @@ class HovmSessionStateKeys(Enum):
     HOVM_END_DATE = "hovm_end_date"
     HOVM_TIMES = "hovm_times"
     HOVM_YAXIS = "hovm_yaxis"
-    HOVM_LEVELS = "hovm_levels"
+    HOVM_P_LEVELS = "hovm_p_levels"
+    HOVM_M_LEVELS = "hovm_m_levels"
     HOVM_DATA_VARIABLE = "hovm_data_variable"
     HOVM_VAR_NAME = "hovm_var_name"
     HOVM_PLOT_TITLE = "hovm_title"
